requirements.txt
django>=3.2,<3.3
django-dotenv
psycopg2-binary
djangorestframework
django-multiupload